midjourney-api/requirements/dev.txt
aiohttp==3.8.4
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.3.0
distlib==0.3.6
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
frozenlist==1.3.3
h11==0.14.0
identify==2.5.24
idna==3.4
iso8601==1.1.0
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
orjson==3.9.1
passlib==1.7.4
platformdirs==3.5.3
pre-commit==3.3.3
pyasn1==0.5.0
pydantic==1.10.9
PyMySQL==1.0.3
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
tortoise-orm==0.19.3
typing_extensions==4.6.3
uvicorn==0.22.0
virtualenv==20.23.0
yarl==1.9.2
